Tested all RS Methods

+++ test/geoplatform-services-test/src/test/java/org/geosdi/geoplatform/model/rest/BasicRestServiceTest.java
@@ -124,12 +124,15 @@ abstract class BasicRestServiceTest extends ServiceTest {
     @Override
     public void tearDown() {
         try {
-            Assert.assertEquals(Boolean.TRUE, gpWSClient.deleteProject(
-                    idProjectTest));
+            if (idProjectTest != -1) {
+                Assert.assertEquals(Boolean.TRUE, gpWSClient.deleteProject(
+                        idProjectTest));
+            }
         } catch (Exception ex) {
-            logger.error("\n@@@@@@@@@@@@@@@@@ERROR@@@@@@@@@@@@@@@@@@@@@@@@ " + ex);
+            logger.error(
+                    "\n@@@@@@@@@@@@@@@@@ERROR@@@@@@@@@@@@@@@@@@@@@@@@ " + ex);
         }
-        
+
         super.tearDown();
     }
